requirements.txt
liboqs-python>=0.10.0
cryptography>=43.0.0
